+++ test/audits/guidanceHighFormViewsLowConversions.test.js
@@ -16,7 +16,7 @@ import { expect, use } from 'chai';
 import sinon from 'sinon';
 import { ok } from '@adobe/spacecat-shared-http-utils';
 import sinonChai from 'sinon-chai';
-import handler from '../../src/forms-opportunities/oppty-handlers/guidance:high-form-views-low-conversions.js';
+import handler from '../../src/forms-opportunities/oppty-handlers/guidance-high-form-views-low-conversions.js';
 
 use(sinonChai);
 
@@ -63,7 +63,8 @@ describe('Guidance High Form Views Low Conversions Handler', () => {
 
   it('should update an existing opportunity', async () => {
     const existingOpportunity = {
-      getData: sinon.stub().returns({ page: 'https://example.com' }),
+      getData: sinon.stub().returns({ form: 'https://example.com' }),
+      getType: sinon.stub().returns('high-page-views-low-conversion'),
       setAuditId: sinon.stub(),
       setGuidance: sinon.stub(),
       save: sinon.stub().resolvesThis(),
@@ -75,7 +76,7 @@ describe('Guidance High Form Views Low Conversions Handler', () => {
     await handler(message, context);
 
     expect(existingOpportunity.setAuditId).to.be.calledWith('audit-id');
-    expect(existingOpportunity.setGuidance).to.be.calledWith({ recommendation: 'Some guidance' });
+    expect(existingOpportunity.setGuidance).to.be.calledWith({ recommendations: 'Some guidance' });
     expect(existingOpportunity.save).to.be.calledOnce;
   });
 
@@ -90,22 +91,6 @@ describe('Guidance High Form Views Low Conversions Handler', () => {
     await handler(message, context);
 
     expect(dataAccessStub.Opportunity.create.callCount).to.equal(0);
-  });
-
-  it('should handle suggestions correctly', async () => {
-    const existingOpportunity = {
-      getData: sinon.stub().returns({ page: 'https://example.com' }),
-      setAuditId: sinon.stub(),
-      setGuidance: sinon.stub(),
-      save: sinon.stub().resolvesThis(),
-      getId: sinon.stub().resolves('testId'),
-      getSuggestions: sinon.stub().resolves([{ remove: sinon.stub().resolves() }]),
-    };
-    dataAccessStub.Opportunity.allBySiteId.resolves([existingOpportunity]);
-
-    await handler(message, context);
-
-    expect(dataAccessStub.Suggestion.create).to.have.been.calledOnce;
   });
 
   it('should return ok response', async () => {